requirements.txt
numpy
pandas
matplotlib
scikit-learn
streamlit
torch
git+https://github.com/openai/CLIP.git
pillow
tqdm
ftfy
regex